refactor(UpdatePlayerRequest): rename to be consistent with User and Game
